200-200-100 Client und Serverkomponenten - connect gui and daemon
Connect xanoubis with anoubisd.
For this the Communicator was changed to 'be' an own thread, this
shall enable us to use the same logic as in the daemon.
In addition arriving notifies will be handed to ModAnoubis.
The NotifyListElement was extended to extract the informatoion
of an arrived notification.

Aegis-Change: anoubis.10.650.C469 Delta 93
Bug: 300
TE BLE

+++ src/libanoubisprotocol/anoubis_dump.c
@@ -86,7 +86,7 @@ static void dump_versel(Anoubis_VerselMessage * m, size_t len __used)
 static void dump_stringlist(Anoubis_StringListMessage * m, size_t len)
 {
 	len -= sizeof(Anoubis_StringListMessage);
-	printf(" stringlist = %.*s", len, m->stringlist);
+	printf(" stringlist = %.*s", (int)len, m->stringlist);
 }
 
 static void dump_ack(Anoubis_AckMessage * m, size_t len __used)
@@ -101,7 +101,7 @@ static void dump_authreply(Anoubis_AuthReplyMessage * m, size_t len)
 	DUMP_NETX(m, error);
 	DUMP_NETU(m, uid);
 	len -= sizeof(Anoubis_AuthReplyMessage);
-	printf(" name = %.*s", len, m->name);
+	printf(" name = %.*s", (int)len, m->name);
 }
 
 static void dump_notify(Anoubis_NotifyMessage * m, size_t len)
